refactor: 프로필 마커 파일 병합 및 위치 변경
