+++ apps/web/src/components/search-place-bottom-sheet/index.tsx
@@ -18,11 +18,12 @@ import CurrentLocationIcon from "../../assets/icons/CurrentLocationIcon";
 import { getLatLng, Marker, NaverLatLng, useNaverMap } from "@repo/naver-map";
 import { convertWGS84ToLatLng, getFullAddressAndTitle } from "@/utils/location";
 import { useCurrentLocation } from "@/hooks/api/useCurrentLocation";
-import ProfileMarker from "./profile-marker";
+
 import { useSelectStartPlace } from "@/hooks/api/useSelectStartPlace";
 import { useRouter } from "next/navigation";
 import Image from "next/image";
 import { useSetStartLocation } from "@/hooks/api/useStartLocation";
+import ProfileMarker from "node_modules/@repo/naver-map/src/overlays/profile-marker";
 
 interface SearchPlaceBottomSheetProps {
   roomId: string;
@@ -49,14 +50,17 @@ const SearchPlaceBottomSheet = ({
   const { data: searchList, refetch: fetchSearchList } =
     useGetPlaceSearchList(query);
   const { setStartLocation } = useSetStartLocation();
+  const profileMarkerElement = ProfileMarker({ profileImageUrl: memberImgUrl });
   const marker = Marker({
     map: map!,
-    customMarkerData: {
-      marker: ProfileMarker({ profileImageUrl: memberImgUrl }),
-      width: 48,
-      height: 48,
-    },
-  }); // TODO: 프로필 URL을 전달받아 렌더링
+    customMarkerData: profileMarkerElement
+      ? {
+          marker: profileMarkerElement,
+          width: 48,
+          height: 48,
+        }
+      : undefined, // NOTE: undefined시 기본 마커 사용
+  });
 
   const moveTo = useCallback(
     (latLng: NaverLatLng) => {
