+++ packages/naver-map/src/overlays/index.ts
@@ -1,1 +1,2 @@
 export * from "./marker";
+export * from "./profile-marker";
